Add DEAP implementation of "complex regression"
The target here is (x^3 + 1)^3 + 1. To get reasonable chances of success, I needed to add `square` and `double` functions as "stand-ins" for Push's `Dup` instruction.

I also needed to cap the value of `f(x)` since it could get really big and blow up Python.

I set the population size to 1K and the max generations to 100, as that tends to give us some reasonable room to work with, and it sometimes even finds a solution.